GA sucks. There. I said it. I've writtent the skeleton all right: - Randomly initialize the population. - Create a mating pool by oversampling the fittest organisms. - Randomly pick 2 for mating. - Do a crossover. - Mutate the child. - Add to new population. - Replace old population with new population. - And repeat until converegence.
THIS DOES NOT WORK!!!

I think the problem is with the `create_pool` function.
After oversampling, my chances of selecting the same organism as
both parents is very likely.

A fitness weighted selection directly from the initial population
should be a lot better.

The 'midpoint' crossover could also be improved, but one thing at
a time.